+++ mangooio-integration-test/src/test/java/io/mangoo/utils/RequestUtilsTest.java
@@ -12,6 +12,7 @@ import io.undertow.server.HttpServerExchange;
 import io.undertow.util.AttachmentKey;
 import io.undertow.util.HeaderMap;
 import io.undertow.util.Methods;
+import io.undertow.util.PathTemplateMatch;
 import org.hamcrest.MatcherAssert;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -263,6 +264,71 @@ class RequestUtilsTest {
         assertThat(parameters.get("queryParam"), equalTo("queryValue"));
         assertThat(parameters.get("pathParam"), equalTo("pathValue"));
         assertThat(parameters.size(), equalTo(2));
+    }
+
+    @Test
+    void testPathParameterWinsOverQueryParameterOfSameName() {
+        //given
+        HttpServerExchange exchange = new HttpServerExchange(null);
+        exchange.addQueryParam("id", "spoofed");
+        exchange.putAttachment(PathTemplateMatch.ATTACHMENT_KEY,
+                new PathTemplateMatch("/foo/{id}", Map.of("id", "routevalue")));
+
+        //then
+        assertThat(RequestUtils.getRequestParameters(exchange).get("id"), equalTo("routevalue"));
+        assertThat(RequestUtils.getPathParameters(exchange).get("id"), equalTo("routevalue"));
+        assertThat(RequestUtils.getQueryParameters(exchange).get("id"), equalTo("spoofed"));
+    }
+
+    @Test
+    void testPathTemplateMatchWinsOverAddPathParam() {
+        //given the router resolved value is authoritative
+        HttpServerExchange exchange = new HttpServerExchange(null);
+        exchange.addPathParam("id", "fromHandler");
+        exchange.putAttachment(PathTemplateMatch.ATTACHMENT_KEY,
+                new PathTemplateMatch("/foo/{id}", Map.of("id", "fromRouter")));
+
+        //then
+        assertThat(RequestUtils.getPathParameters(exchange).get("id"), equalTo("fromRouter"));
+    }
+
+    @Test
+    void testGetQueryParametersHoldsNoRouteValues() {
+        //given
+        HttpServerExchange exchange = new HttpServerExchange(null);
+        exchange.addQueryParam("limit", "25");
+        exchange.putAttachment(PathTemplateMatch.ATTACHMENT_KEY,
+                new PathTemplateMatch("/foo/{id}", Map.of("id", "routevalue")));
+
+        //when
+        Map<String, String> queryParameters = RequestUtils.getQueryParameters(exchange);
+
+        //then
+        assertThat(queryParameters.get("limit"), equalTo("25"));
+        assertThat(queryParameters.containsKey("id"), equalTo(false));
+        assertThat(queryParameters.size(), equalTo(1));
+    }
+
+    @Test
+    void testHasAmbiguousParameters() {
+        //given
+        HttpServerExchange colliding = new HttpServerExchange(null);
+        colliding.addQueryParam("id", "spoofed");
+        colliding.putAttachment(PathTemplateMatch.ATTACHMENT_KEY,
+                new PathTemplateMatch("/foo/{id}", Map.of("id", "routevalue")));
+
+        HttpServerExchange distinct = new HttpServerExchange(null);
+        distinct.addQueryParam("limit", "25");
+        distinct.putAttachment(PathTemplateMatch.ATTACHMENT_KEY,
+                new PathTemplateMatch("/foo/{id}", Map.of("id", "routevalue")));
+
+        HttpServerExchange withoutTemplate = new HttpServerExchange(null);
+        withoutTemplate.addQueryParam("id", "anything");
+
+        //then
+        assertThat(RequestUtils.hasAmbiguousParameters(colliding), equalTo(true));
+        assertThat(RequestUtils.hasAmbiguousParameters(distinct), equalTo(false));
+        assertThat(RequestUtils.hasAmbiguousParameters(withoutTemplate), equalTo(false));
     }
 
     @Test